+++ ivadomed/utils.py
@@ -9,7 +9,6 @@ import torch
 import torch.nn as nn
 
 
-
 AXIS_DCT = {'sagittal': 0, 'coronal': 1, 'axial': 2}
 
 # List of classification models (ie not segmentation output)
@@ -41,58 +40,6 @@ def cuda(input_var, cuda_available=True, non_blocking=False):
             return input_var.cuda(non_blocking=non_blocking)
     else:
         return input_var
-
-
-
-
-class SliceFilter(object):
-    """Filter 2D slices from dataset.
-
-    If a sample does not meet certain conditions, it is discarded from the dataset.
-
-    Args:
-        filter_empty_mask (bool): If True, samples where all voxel labels are zeros are discarded.
-        filter_empty_input (bool): If True, samples where all voxel intensities are zeros are discarded.
-
-    Attributes:
-        filter_empty_mask (bool): If True, samples where all voxel labels are zeros are discarded.
-        filter_empty_input (bool): If True, samples where all voxel intensities are zeros are discarded.
-    """
-
-    def __init__(self, filter_empty_mask=True,
-                 filter_empty_input=True,
-                 filter_classification=False, classifier_path=None, device=None, cuda_available=None):
-        self.filter_empty_mask = filter_empty_mask
-        self.filter_empty_input = filter_empty_input
-        self.filter_classification = filter_classification
-        self.device = device
-        self.cuda_available = cuda_available
-
-        if self.filter_classification:
-            if cuda_available:
-                self.classifier = torch.load(classifier_path, map_location=device)
-            else:
-                self.classifier = torch.load(classifier_path, map_location='cpu')
-
-    def __call__(self, sample):
-        input_data, gt_data = sample['input'], sample['gt']
-
-        if self.filter_empty_mask:
-            if not np.any(gt_data):
-                return False
-
-        if self.filter_empty_input:
-            # Filter set of images if one of them is empty or filled with constant value (i.e. std == 0)
-            if np.any([img.std() == 0 for img in input_data]):
-                return False
-
-        if self.filter_classification:
-            if not np.all([int(
-                    self.classifier(cuda(torch.from_numpy(img.copy()).unsqueeze(0).unsqueeze(0), self.cuda_available)))
-                           for img in input_data]):
-                return False
-
-        return True
 
 
 def unstack_tensors(sample):
